create reusable trip details widget
