+++ lib/widgets/trip_details.dart
@@ -1,0 +1,135 @@
+import 'package:flutter/material.dart';
+import 'package:go_ride_sharing/models/post.dart';
+import 'package:go_ride_sharing/models/conversation.dart';
+import 'package:go_ride_sharing/pages/conversation_detail_page.dart';
+import 'package:go_ride_sharing/services/post_service.dart';
+import 'package:go_ride_sharing/services/message_service.dart';
+import 'package:go_ride_sharing/theme.dart';
+
+class TripDetails extends StatelessWidget {
+  final Post post;
+
+  const TripDetails({super.key, required this.post});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        RichText(
+          text: TextSpan(
+            text: 'Origin: ',
+            style: const TextStyle(
+              fontSize: 14.0,
+              fontWeight: FontWeight.bold,
+              color: Colors.black,
+            ),
+            children: [
+              TextSpan(
+                text: post.originName,
+                style: const TextStyle(
+                  fontWeight: FontWeight.normal,
+                ),
+              ),
+            ],
+          ),
+        ),
+        const SizedBox(height: 5.0),
+        RichText(
+          text: TextSpan(
+            text: 'Destination: ',
+            style: const TextStyle(
+              fontSize: 14.0,
+              fontWeight: FontWeight.bold,
+              color: Colors.black,
+            ),
+            children: [
+              TextSpan(
+                text: post.destinationName,
+                style: const TextStyle(
+                  fontWeight: FontWeight.normal,
+                ),
+              ),
+            ],
+          ),
+        ),
+        const SizedBox(height: 5.0),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            RichText(
+              text: TextSpan(
+                text: 'Date: ',
+                style: const TextStyle(
+                  fontSize: 14.0,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.black,
+                ),
+                children: [
+                  TextSpan(
+                    text: post.departureDate.toLocal().toString().split(' ')[0],
+                    style: const TextStyle(
+                      fontSize: 14.0,
+                      fontWeight: FontWeight.normal,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 5.0),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              'Max ${post.seatsAvailable} passengers',
+              style: const TextStyle(
+                fontSize: 14.0,
+              ),
+            ),
+            Text(
+              '\$${post.price.toStringAsFixed(2)}',
+              style: const TextStyle(
+                fontSize: 14.0,
+              ),
+            ),
+          ],
+        ),
+        const SizedBox(height: 10.0),
+        FutureBuilder<String?>(
+          future: PostService().userID,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return CircularProgressIndicator();
+            } else if (snapshot.hasData && snapshot.data != post.posterId) {
+              return ElevatedButton(
+                onPressed: () async {
+                  Conversation conversation =
+                      await MessageService().createConversation(post.posterId);
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) =>
+                          ConversationDetailPage(conversation: conversation),
+                    ),
+                  );
+                },
+                style: FilledButton.styleFrom(
+                    backgroundColor: notYellow,
+                    foregroundColor: notBlack,
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(10),
+                    ),
+                    shadowColor: notBlack),
+                child: const Text('Contact'),
+              );
+            } else {
+              return const SizedBox.shrink();
+            }
+          },
+        ),
+      ],
+    );
+  }
+}
